# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: The Champions Club API
  version: 1.0.0
  description: "GENERATED from tools/api/endpoints.mjs — DO NOT EDIT BY HAND. Human-readable version: API_CONTRACT.md. Schemas are derived from shared/types/*.ts (TypeScript is authoritative; inline/complex shapes may appear as free-form)."
servers:
  - url: /api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      properties:
        success:
          description: "TS: false"
        error:
          type: object
          properties:
            code:
              description: "TS: ErrorCode"
            message:
              type: string
            details:
              type: object
              additionalProperties: {}
          required:
            - code
            - message
      required:
        - success
        - error
...
    Staff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        designation:
          type: string
        monthly_salary:
          type: string
          pattern: ^\d+\.\d{2}$
          example: "1250.00"
        joined_on:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - designation
        - monthly_salary
        - joined_on
        - created_at
        - updated_at
    BusinessClient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
          nullable: true
        company_name:
          type: string
        contact_name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        gstin:
          type: string
          nullable: true
        billing_address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - user_id
        - company_name
        - contact_name
        - email
        - phone
        - gstin
        - billing_address
        - notes
        - is_active
        - created_at
        - updated_at
...
    AuthSession:
      type: object
      properties:
        token:
          type: string
        expires_at:
          type: string
          format: date-time
        user:
          description: "TS: Omit<User, 'password_hash'>"
        redirect_to:
          type: string
        member:
          $ref: "#/components/schemas/MemberSummary"
          nullable: true
        staff:
          $ref: "#/components/schemas/StaffView"
          nullable: true
        business_client:
          $ref: "#/components/schemas/BusinessClient"
          nullable: true
      required:
        - token
        - expires_at
        - user
        - redirect_to
        - member
        - staff
        - business_client
    ActiveMembershipSummary:
      type: object
      properties:
        membership_id:
          type: string
          format: uuid
        membership_plan_id:
          type: string
          format: uuid
        membership_type:
          type: string
          enum:
            - GOLD
            - SILVER
            - JUNIOR
        plan_name:
          type: string
        status:
          type: string
          enum:
            - UPCOMING
            - ACTIVE
            - EXPIRED
            - CANCELLED
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        days_remaining:
          type: number
      required:
        - membership_id
        - membership_plan_id
        - membership_type
        - plan_name
        - status
        - start_date
        - end_date
        - days_remaining
    MemberSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        member_code:
          type: string
        full_name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        date_of_birth:
          type: string
          format: date
          nullable: true
        photo_url:
          type: string
          nullable: true
        joined_on:
          type: string
          format: date
        is_active:
          type: boolean
        active_membership:
          $ref: "#/components/schemas/ActiveMembershipSummary"
          nullable: true
      required:
        - id
        - user_id
        - member_code
        - full_name
        - email
        - phone
        - date_of_birth
        - photo_url
        - joined_on
        - is_active
        - active_membership
...
    StaffView:
      allOf:
        - $ref: "#/components/schemas/Staff"
        - type: object
          properties:
            full_name:
              type: string
            email:
              type: string
            phone:
              type: string
              nullable: true
            role:
              type: string
              enum:
                - MEMBER
                - FRONT_DESK
                - KITCHEN_MANAGER
                - BUSINESS_CLIENT
                - OWNER_ADMIN
            is_active:
              type: boolean
          required:
            - full_name
            - email
            - phone
            - role
            - is_active
...
paths:
  /auth/signup:
    post:
      operationId: auth.signup
      tags:
        - Authentication
      summary: Visitor creates an account (role MEMBER, no plan yet) and is logged in.
      description: |-
        Creates `users` (role MEMBER) + `members` in one transaction. No membership is created: until a plan is bought the member pays walk-in rates.

        Requirements: FR-AUTH-001
      x-roles:
        - PUBLIC
      x-owner: Dev 1
      x-requirements:
        - FR-AUTH-001
      x-tables:
        - users
        - members
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  description: 2-120 chars
                email:
                  type: string
                  format: email
                  description: unique (case-insensitive)
                phone:
                  type: string
                password:
                  type: string
                  description: min 8 chars, >=1 letter and >=1 digit
                date_of_birth:
                  type: string
                  format: date
                  description: must be in the past
              required:
                - full_name
                - email
                - phone
                - password
      responses:
        "201":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/AuthSession"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: EMAIL_TAKEN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/login:
    post:
      operationId: auth.login
      tags:
        - Authentication
      summary: Authenticate and receive a JWT plus the role-based landing route.
      description: |-
        `redirect_to` = ROLE_HOME_ROUTE[user.role]. Never reveal whether the email exists (always AUTH_INVALID).

        Requirements: FR-AUTH-002, FR-AUTH-003
      x-roles:
        - PUBLIC
      x-owner: Dev 1
      x-requirements:
        - FR-AUTH-002
        - FR-AUTH-003
      x-tables:
        - users
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required:
                - email
                - password
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/AuthSession"
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: ACCOUNT_DISABLED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/logout:
    post:
      operationId: auth.logout
      tags:
        - Authentication
      summary: "End the session (stateless JWT: the client discards the token)."
      description: "Requirements: FR-AUTH-004"
      x-roles:
        - MEMBER
        - FRONT_DESK
        - KITCHEN_MANAGER
        - BUSINESS_CLIENT
        - OWNER_ADMIN
      x-owner: Dev 1
      x-requirements:
        - FR-AUTH-004
      x-tables: []
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    nullable: true
                  message:
                    type: string
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/me:
    get:
      operationId: auth.me
      tags:
        - Authentication
      summary: Current user + role profile + landing route. Re-issues a fresh token.
      description: "Requirements: FR-AUTH-002"
      x-roles:
        - MEMBER
        - FRONT_DESK
        - KITCHEN_MANAGER
        - BUSINESS_CLIENT
        - OWNER_ADMIN
      x-owner: Dev 1
      x-requirements:
        - FR-AUTH-002
      x-tables:
        - users
        - members
        - staff
        - business_clients
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    $ref: "#/components/schemas/AuthSession"
                  message:
                    type: string
        "401":
          description: AUTH_UNAUTHORIZED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/change-password:
    post:
      operationId: auth.changePassword
      tags:
        - Authentication
      summary: Change own password (clears must_change_password).
      description: "Requirements: FR-AUTH-006"
      x-roles:
        - MEMBER
        - FRONT_DESK
        - KITCHEN_MANAGER
        - BUSINESS_CLIENT
        - OWNER_ADMIN
      x-owner: Dev 1
      x-requirements:
        - FR-AUTH-006
      x-tables:
        - users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  description: min 8 chars, >=1 letter and >=1 digit
              required:
                - current_password
                - new_password
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - data
                properties:
                  success:
                    type: boolean
                    enum:
                      - true
                  data:
                    nullable: true
                  message:
                    type: string
        "400":
          description: VALIDATION_ERROR
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: AUTH_UNAUTHORIZED | AUTH_INVALID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...